+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 6.1.0
 -----
 
-* Added claim support for proctoring end assessment message
+* Added claim support for [proctoring end assessment message](https://www.imsglobal.org/spec/proctoring/v1p0#h.ooq616k28cwm)
 
 6.0.1
 -----